fix(bloomery): avoid From<u8> in const verifier-set helpers
const fn cannot call usize::from on u8 on this toolchain; compare the sidecar length against the slot bound as u8 instead.

+++ crates/aether-bloomery/src/values/verify.rs
@@ -639,7 +639,7 @@ impl VerifyFailureSet {
     /// Insert `identity` in position order, or keep the existing occupant of
     /// that bit.
     const fn push_declared(mut self, identity: DeclaredIdentity) -> Self {
-        if usize::from(self.declared_len) >= DECLARED_SLOTS {
+        if self.declared_len >= DECLARED_SLOTS as u8 {
             return self;
         }
         let mut index = 0;
